make some api private
